Add function 'operate'
To perform a math operation based on the user's input, the function
'operate' was implemented. It will output a number based on the
numbers the user has inputted, as well as the math operator that was
selected.

Additionally, three variables have been added to store and display
the user's inputted numbers and math operator.