Add updateCalcDisplay function
To update the number being displayed as the user clicks on different
buttons, the updateCalcDisplay function was added that updates the
textContent of the div with the inputted number.

Alongside this, the variable 'calcDisplay' was made to store the
current number being displayed, as well as the 'screenArea' variable
to select the '.screen-area' div in order to insert the current number
into its textContent.