Add changeCalcNumber function
To change the number being displayed on the screen based on what number
the user has inputted, the changeCalcNumber function was created and
added into the click event listener for 'buttonsArea'.

Additionally, a const 'MAX_CALC_NUM_LENGTH' was added for the
aforementioned function in order to prevent very big numbers from being
calculated.