Refactor attributes on calculator buttons
Currently, the script extracts the number of the clicked button based
on its id, which does not make sense given the button has a value
attribute that can be used to obtain the number instead. Thus, majority
of the HTML buttons had their id values extracted and inserted into a
value attribute instead, along with their ids being renamed.

In turn, the script was also refactored to now get the number of the
clicked button based on the button's value rather than their id.