Modify 'operate' method to round decimal numbers
Previously, when getting decimal numbers as a result, the number itself
could overflow outside of the display of the calculator screen. To fix
this, the 'operate' method for the 'calculation' object now includes
logic to round the number based on the 'MAX_NUM_LENGTH' value.

Additionally, the 'operate' method will also return 'NaN' if the result
of the calculation is bigger than 'MAX_NUM_LENGTH', to prevent the
aforementioned issue from occurring.