Add 'highlightBtn' function, 'mathButtons' object
To show what math operator was selected, the 'highlightBtn'
function was added to change the background colour of the math operator
button that was clicked.

For this to work, a new object called 'mathButtons' was introduced that
holds a list of button elements with respect to their math operator.

Additionally, the 'if' statement that checks if the 'input' value was
one of the math operator buttons, inside of the click event listener,
was refactored. Instead of checking if the 'input' value exists within
the 'mathOperators' array, it now checks if said value is an existing
key in the 'mathButtons' object. Consequently, the 'mathOperators'
array was removed as it no longer served its purpose.